Decouple tests and engine.
Now all tests can be built at the same time, without changing defines.
This only fixes the CMake files, VS2010 and Xcode have to follow.
