+++ CMakeLists.txt
@@ -14,7 +14,6 @@ set(LIBRARY_OUTPUT_PATH ${PROJECT_BINARY_DIR}/lib)
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
 
 set(BUILD_TESTS OFF CACHE BOOL "Build tests")
-set(TEST_TO_BUILD "Jumper" CACHE STRING "The test which will be built")
 set(WITH_OPENAL ON CACHE BOOL "Use OpenAL")
 set(WITH_OPENGL ON CACHE BOOL "Use OpenGL")
