+++ Code/Tests/Jumper/CMakeLists.txt
@@ -1,8 +1,8 @@
-set(PONG_SRCS
-	Jumper.cpp
+set(JUMPER_SRCS
+	Main.cpp Jumper.cpp
 )
 
-add_executable(TestJumper ${PONG_SRCS})
+add_executable(TestJumper ${JUMPER_SRCS})
 
 target_link_libraries(TestJumper MonocleCore ${CORE_LINK})
