+++ Code/Tests/Marian/CMakeLists.txt
@@ -1,0 +1,8 @@
+set(MARIAN_SRCS
+	Main.cpp Marian.cpp
+)
+
+add_executable(TestMarian ${MARIAN_SRCS})
+
+target_link_libraries(TestMarian MonocleCore ${CORE_LINK})
+
