+++ Code/Tests/Pong/Main.cpp
@@ -1,0 +1,14 @@
+#include "Monocle.h"
+#include "Pong.h"
+
+using namespace Monocle;
+
+int main(void)
+{
+	Game *game = new Game();
+	game->Init();
+	game->SetScene(new Pong::GameScene());
+	game->Main();
+	delete game;
+	return 0;
+}
